+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - #task_number - short task description
 - #2 - entities personal_address, pickup_point, postomat where created, integrated into gazon-service and tested
+- #4 - entity BankCard was created, integrated into gazon-service and tested
 
 ## [1.0.0]